DRY up archive extraction helper methods
