+++ app/jobs/process_uploaded_file_job.rb
@@ -1,4 +1,6 @@
 class ProcessUploadedFileJob < ApplicationJob
+  include ArchiveHelpers
+
   queue_as :critical
 
   def perform(library_id, uploaded_file, name: nil, owner: nil, creator_id: nil, collection_ids: [], tag_list: nil, license: nil, model: nil, sensitive: nil, permission_preset: nil)
@@ -109,28 +111,5 @@ class ProcessUploadedFileJob < ApplicationJob
       end
     end
     new_files
-  end
-
-  def count_common_path_components(archive)
-    # Generate full list of directories in the archive
-    paths = []
-    files_in_root = false
-    Archive::Reader.open_filename(archive.path) do |reader|
-      reader.each_entry do |entry|
-        paths << entry.pathname if entry.directory?
-        files_in_root = true if entry.file? && entry.pathname.exclude?(File::SEPARATOR)
-      end
-    end
-    return 0 if files_in_root
-    paths = paths.map { |path| path.split(File::SEPARATOR) }
-    # Count the common elements in the paths
-    count_common_elements(paths)
-  end
-
-  def count_common_elements(arrays)
-    return 0 if arrays.empty?
-    first = arrays.shift
-    zip = first.zip(*arrays)
-    zip.count { it.uniq.count == 1 }
   end
 end
