+++ spec/jobs/process_uploaded_file_job_spec.rb
@@ -3,49 +3,6 @@ require "rails_helper"
 RSpec.describe ProcessUploadedFileJob do
   subject(:job) { described_class.new }
 
-  context "when counting common path prefixes" do
-    it "returns zero if there are no directories at all" do
-      expect(job.send(:count_common_elements, [])).to eq 0
-    end
-
-    it "returns zero if there are no common prefixes" do
-      expect(job.send(:count_common_elements, [
-        ["folder1"],
-        ["folder2"],
-        []
-      ])).to eq 0
-    end
-
-    it "returns the number of common prefixes if present" do
-      expect(job.send(:count_common_elements, [
-        ["root", "sub", "folder1"],
-        ["root", "sub", "folder2"]
-      ])).to eq 2
-    end
-
-    it "returns correct count where all prefixes are the same" do
-      expect(job.send(:count_common_elements, [
-        ["root", "sub", "folder"],
-        ["root", "sub", "folder"]
-      ])).to eq 3
-    end
-
-    it "returns correct count where a common folder has a subfolder" do
-      expect(job.send(:count_common_elements, [
-        ["root", "sub"],
-        ["root", "sub", "folder2"]
-      ])).to eq 2
-    end
-
-    it "returns zero for *some* common prefixes but not on everything" do
-      expect(job.send(:count_common_elements, [
-        ["folder1", "sub1"],
-        ["folder1", "sub2"],
-        ["folder2", "sub1"]
-      ])).to eq 0
-    end
-  end
-
   context "when uploading a file", :after_first_run do
     let(:uploader) { create(:contributor) }
     let(:library) { create(:library) }
